requirements.txt
youtube_dl
pydub
spleeter
SpeechRecognition
tqdm
webvtt-py